requirements.txt
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
PyYAML==6.0.1
Pillow==10.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

